+++ test-data/src/main/resources/org/openmrs/module/fhir2/providers/Location_json_patch.json
@@ -1,0 +1,7 @@
+[
+  {
+    "op": "replace",
+    "path": "/name",
+    "value": "Patched Location"
+  }
+]
